chore: move pkg/utilities
moved pkg/utilities to internal/utilities to line up with go standards
for project layout

resolves #73
