+++ internal/curseforgeOpener.go
@@ -2,10 +2,11 @@ package internal
 
 import (
 	"fmt"
-	"github.com/spf13/viper"
 	"log"
 	"os/exec"
-	"wowtools/pkg/utilities"
+	"wowtools/internal/utilities"
+
+	"github.com/spf13/viper"
 )
 
 func OpenCurseforge() {
