+++ internal/utilities/blizzardApiHelper.go
