+++ internal/utilities/confirmationHelper.go
