+++ internal/utilities/discordHelper.go
