+++ internal/utilities/logHelper.go
